Review/suggestion fixes
