+++ Dusts/WaterSplash/SulphuricSplash.cs
@@ -3,18 +3,7 @@ using Terraria.ModLoader;
 
 namespace CalamityMod.Dusts.WaterSplash
 {
-    public class SulphuricSplash : ModDust
+    public class SulphuricSplash : SplashDust
     {
-        public override void SetStaticDefaults()
-        {
-            UpdateType = 33;
-        }
-
-        public override void OnSpawn(Dust dust)
-        {
-            dust.alpha = 170;
-            dust.velocity *= 0.5f;
-            dust.velocity.Y += 1f;
-        }
     }
 }
